pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "awq"
version = "0.1.0"
description = "An efficient and accurate low-bit weight quantization(INT3/4) method for LLMs."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "accelerate", "sentencepiece", "tokenizers>=0.12.1",
    "torchvision",
    "transformers>=4.31.0",
    "lm_eval", "texttable",
    "toml", "attributedict",
    "protobuf"
]

[tool.setuptools.packages.find]
exclude = ["results*", "scripts*", "examples*"]

[tool.wheel]
exclude = ["results*", "scripts*", "examples*"]
